Covid-19/requirements.txt
fastapi
uvicorn
pillow
numpy
tensorflow
python-multipart
flask-cors